Draft `google-java-format` aspect
In addition to adding the aspect itself, this commit adds a few
supporting definitions:

- The `google_java_format_toolchain` rule.
- The `google_java_format_toolchain` toolchain type.
- The `remote_google_java_format_repository` repository rule.

This also adds a new Bazel test workspace for which configures and uses
this new aspect: `test/workspaces/smoke_test_google_java_format_aspect`.
This workspace includes two Java targets:

- `@smoke_test_google_java_format_aspect//:BadlyFormatted`
- `@smoke_test_google_java_format_aspect//:WellFormatted`

The former target fails to `bazel build` because `google-java-format`
finds that it needs formatting. (It is 4-space indented, not 2-space
indented.)