Split `CustomJavaInfo` into two providers
The two replacement providers are `JavaCompilationInfo` and
`JavaDependencyInfo`. This change impacted the implementations of all
Java rules.

This change should pave the way for using legacy Java targets as
dependencies.

I considered using `JavaInfo` directly, but in my initial attempts, it
seemed too hard to work with.

Maybe the biggest change was to the helper function which compiles and
JARs Java source code. Previously, this function was named
`build_jar_from_java_sources()`, but while rewriting it it was renamed
to `compile_and_jar_java_sources()`.

I also renamed and reorganized some other helper functions under
`java/rules/common`.

Note that the `java_import` rule now has a `jars` attribute. (It was
previously just a `jar`.)