Draft `java_agent`; add attr for JAR manifest attr
A Java agent can be used in [a few different ways][1]. One way requires
a JAR with a special manifest attibute, `Premain-Class`. This commit
adds a `java_agent` macro with which one can create such a JAR.

To implement this, this commit also adds the `include_in_jar_manifest`
attribute to the `java_test`, `java_library`, and `java_binary` rules.
This rule attribute is meant to be used to add arbitrary attributes to
JAR manifests.

Note that these Java rules do not yet provide a way to actually *use* a
Java agent JAR. (That is, there not yet a way to add a `-javaagent`
argument to a `java` invocation.)

This commit also adds a new test workspace to help test this new macro:
`test/workspaces/smoke_test_java_agent_rule`.

---

1: https://docs.oracle.com/javase/7/docs/api/java/lang/instrument/package-summary.html