Draft Error Prone aspect
This introduces a number of definitions:

- The `error_prone_aspect` aspect. When enabled, this re-compiles all
  Java targets with the Error Prone `javac` plugin.
- The `error_prone_toolchain:toolchain_type`. The above aspect depends
  upon an instance of this to find a copy of the Error Prone `javac`
  plugin.
- The `error_prone_toolchain` rule. This is used to define toolchain
  instances of the above type.
- The `maven_error_prone_repository` repository rule. This uses Maven
  (via `@rules_jvm_external`) to fetch a copy of Error Prone and to wrap
  it in an `error_prone_toolchain`.
- The `smoke_test_error_prone_aspect` test workspace. This tests these
  new definitions.

Warning: Error Prone warnings are currently logged but they do not cause
build errors and they are not printed to the `bazel` client's
stdout/stderr. This is surprising and should be fixed.

Also tweaked a few comments.

Also fixed possible source of bugs: `File.short_path` was used a few
places where `File.path` was more appropriate.