Remove unneeded `@dwtj_rules_java` from load stmts
It turns out that these prefixes are not needed. This was done to
attempt to prepare support for `stardoc`. At first glance, it seems that
`stardoc` doesn't follow these load statements correctly when they have
a hard-coded `@dwtj_rules_java`. This is probably a `stardoc` bug, but
I'll just work around it. This style is clearer anyway.